@@ -2,11 +2,9 @@ package metrics
 
 import (
 	"context"
-	"fmt"
 	"time"
 
 	log "github.com/sirupsen/logrus"
-
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric"
@@ -40,7 +38,7 @@ func WithMetricsMultiAddress(multiAddrs []MultiAddrs) error {
 	// Define the callback function that will be called periodically to observe metrics.
 	callback := func(ctx context.Context, observer metric.Observer) error {
 		for _, ma := range multiAddrs {
-			// Create labels with attributes for each multiaddress.
+			// Create labels with attributes for each multi-address.
 			labels := metric.WithAttributes(
 				attribute.String("service_name", ma.ServiceName),
 				attribute.String("node_name", ma.NodeName),
@@ -61,9 +59,9 @@ func WithMetricsMultiAddress(multiAddrs []MultiAddrs) error {
 
 // BlockHeight represents the information for the block height 1.
 type BlockHeight struct {
-	ServiceName string  // ServiceName Name of the service associated with the multiaddress.
-	BlockHeight string  // Namespace where the service is deployed.
-	Value       float64 // Value to be observed for the multiaddress.
+	ServiceName string  // ServiceName Name of the service associated with the multi-address.
+	BlockHeight string  // BlockHeight height of the block.
+	Value       float64 // Value to be observed for the multi-address.
 }
 
 // WithMetricsBlockHeight creates a callback function to observe metrics for block_height_1.
@@ -100,11 +98,12 @@ func WithMetricsBlockHeight(blockHeight, earliestBlockTime, serviceName, namespa
 	return err
 }
 
+// CalculateDaysDifference based on the date received, returns the number of days since this day.
 func CalculateDaysDifference(inputTimeString string) int {
 	layout := "2006-01-02T15:04:05.999999999Z"
 	inputTime, err := time.Parse(layout, inputTimeString)
 	if err != nil {
-		fmt.Println("Error parsing time:", err)
+		log.Error("Error parsing time: [", inputTimeString, "]", err)
 		return -1
 	}
 
